+++ src/editors/sampledvalues/subscriber-ied-list-smv.ts
@@ -23,13 +23,12 @@ import {
 import {
   SmvSelectEvent,
   IEDSampledValuesSubscriptionEvent,
-  styles,
 } from './foundation.js';
 import {
   emptyInputsDeleteActions,
   getFcdaReferences
 } from "../../foundation/ied.js";
-import { SubscribeStatus } from '../foundation.js';
+import { styles, SubscribeStatus } from '../foundation.js';
 
 /**
  * An IED within this IED list has 2 properties:
